pyproject.toml
[tool.poetry]
name = "dgradui"
version = "0.1.0"
description = ""
authors = ["Daniel Winkler <dan@danwink.com>"]

[tool.poetry.dependencies]
python = "^3.10"
gradio = "^3.5"
requests = "^2.28.1"
Pillow = "^9.2.0"
opencv-python = "^4.6.0"
diskcache = "^5.4.0"
pycsg = "^0.3.3"
trimesh = "^3.15.5"
shapely = "^1.8.5"
mapbox-earcut = "^1.0.0"
tqdm = "^4.64.1"
solidpython = "^1.1.3"

[tool.poetry.dev-dependencies]
black = "^22.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
